requirements.txt
requests
python-dotenv
numpy
pandas
anthropic
websocket-client
krakenex
httpx
lightgbm>=4.3.0
scikit-learn>=1.4.0
pyarrow>=15.0.0
tqdm>=4.66.0
